Removed storage dependency from eventhubs (#4954)
* Removed storage dependency from eventhubs

---------
