+++ cmake-modules/FolderList.cmake
@@ -33,6 +33,10 @@ macro(GetFolderList project)
     elseif(${project} STREQUAL EVENTHUBS)
         DownloadDepVersion(sdk/core azure-core 1.10.1)
         DownloadDepVersion(sdk/core azure-core-amqp 1.0.0-beta.1)
+    elseif(${project} STREQUAL EVENTHUBS_CHECKPOINT_BLOB)
+        DownloadDepVersion(sdk/core azure-core 1.10.1)
+        DownloadDepVersion(sdk/core azure-core-amqp 1.0.0-beta.1)
+        DownloadDepVersion(sdk/eventhubs azure-messaging-eventhubs 1.0.0-beta.3)
         DownloadDepVersion(sdk/storage/azure-storage-common azure-storage-common 12.3.3)
         DownloadDepVersion(sdk/storage/azure-storage-blobs azure-storage-blobs 12.8.0)
     endif()
