+++ sdk/eventhubs/assets.json
@@ -2,5 +2,5 @@
   "AssetsRepo": "Azure/azure-sdk-assets",
   "AssetsRepoPrefixPath": "cpp",
   "TagPrefix": "cpp/eventhubs",
-  "Tag": "cpp/eventhubs_ea4655bf2e"
+  "Tag": "cpp/eventhubs_d59a0a9f3c"
 }
