+++ sdk/eventhubs/azure-messaging-eventhubs/vcpkg/Config.cmake.in
@@ -5,7 +5,6 @@
 
 include(CMakeFindDependencyMacro)
 find_dependency(azure-core-amqp-cpp)
-find_dependency(azure-storage-blobs-cpp)
 
 include("${CMAKE_CURRENT_LIST_DIR}/azure-messaging-eventhubs-cppTargets.cmake")
