+++ sdk/eventhubs/azure-messaging-eventhubs/vcpkg/vcpkg.json
@@ -17,11 +17,6 @@
       "version>=": "1.0.0-beta.2"
     },
     {
-      "name": "azure-storage-blobs-cpp",
-      "default-features": false,
-      "version>=": "12.8.0"
-    },
-    {
       "name": "vcpkg-cmake",
       "host": true
     },
