+++ sdk/eventhubs/blob-store/CHANGELOG.md
@@ -1,0 +1,13 @@
+# Release History
+
+## 1.0.0-beta.3 (Unreleased)
+
+### Features Added
+
+- Initial release. Split from the `azure-messaging-eventubs` package.
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
