+++ sdk/eventhubs/blob-store/test/blob_checkpoint_store_test.cpp
@@ -1,19 +1,21 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+#include "azure/messaging/eventhubs/checkpointstore_blob/blob_checkpoint_store.hpp"
 #include "eventhubs_test_base.hpp"
-#include "test_checkpoint_store.hpp"
 
 #include <azure/core/context.hpp>
 #include <azure/identity.hpp>
 #include <azure/messaging/eventhubs.hpp>
 
-#include <gtest/gtest.h>
-
 namespace Azure { namespace Messaging { namespace EventHubs { namespace Test {
 
-  class CheckpointStoreTest : public EventHubsTestBase {
-    virtual void SetUp() override { EventHubsTestBase::SetUp(); }
+  class BlobCheckpointStoreTest : public EventHubsTestBase {
+    virtual void SetUp() override
+    {
+      EventHubsTestBase::SetUp();
+      m_blobClientOptions = InitClientOptions<Azure::Storage::Blobs::BlobClientOptions>();
+    }
 
   protected:
     std::string GetRandomName()
@@ -29,13 +31,17 @@ namespace Azure { namespace Messaging { namespace EventHubs { namespace Test {
       }
       return name;
     }
+
+    Azure::Storage::Blobs::BlobClientOptions m_blobClientOptions;
   };
 
-  TEST_F(CheckpointStoreTest, TestCheckpoints)
+  TEST_F(BlobCheckpointStoreTest, TestCheckpoints)
   {
     std::string const testName = GetRandomName();
     std::string consumerGroup = GetEnv("EVENTHUB_CONSUMER_GROUP");
-    Azure::Messaging::EventHubs::Test::TestCheckpointStore checkpointStore;
+    auto containerClient{Azure::Storage::Blobs::BlobContainerClient::CreateFromConnectionString(
+        GetEnv("CHECKPOINTSTORE_STORAGE_CONNECTION_STRING"), testName, m_blobClientOptions)};
+    Azure::Messaging::EventHubs::BlobCheckpointStore checkpointStore(containerClient);
 
     auto checkpoints = checkpointStore.ListCheckpoints(
         "fully-qualified-namespace", "event-hub-name", "consumer-group");
@@ -81,11 +87,13 @@ namespace Azure { namespace Messaging { namespace EventHubs { namespace Test {
     EXPECT_EQ(102, checkpoints[0].Offset.Value());
   }
 
-  TEST_F(CheckpointStoreTest, TestOwnerships)
+  TEST_F(BlobCheckpointStoreTest, TestOwnerships)
   {
     std::string const testName = GetRandomName();
+    auto containerClient{Azure::Storage::Blobs::BlobContainerClient::CreateFromConnectionString(
+        GetEnv("CHECKPOINTSTORE_STORAGE_CONNECTION_STRING"), testName, m_blobClientOptions)};
 
-    TestCheckpointStore checkpointStore;
+    Azure::Messaging::EventHubs::BlobCheckpointStore checkpointStore(containerClient);
 
     auto ownerships = checkpointStore.ListOwnership(
         "fully-qualified-namespace", "event-hub-name", "consumer-group");
