+++ vcpkg.json
@@ -1,7 +1,7 @@
 {
   "name": "azure-sdk-for-cpp",
   "version": "1.5.0",
-  "builtin-baseline": "dafef74af53669ef1cc9015f55e0ce809ead62aa",
+  "builtin-baseline": "33409307f1e3411112a0a6bbf3011ea3cca1bfc9",
   "dependencies": [
     {
       "name": "curl"
